move methods to utils class
